ajustes

+++ back-end/app.js
@@ -8,10 +8,10 @@ const crypto = require('crypto');
 const util = require('util')
 const randomBytes = util.promisify(crypto.randomBytes)
 const emailTemplate = require('./email/template')
-const TokenVerify = require('./models/TokenVerify')
-const Token = require('./models/Token')
 require('dotenv').config()
 const loginAuth = require('./auth/login')
+const jwt = require("jsonwebtoken");
+const JWTSecret = process.env["JWT_SECRET"];
 
 // API config
 app.use(bodyParser.urlencoded({
@@ -22,27 +22,26 @@ app.use(bodyParser.json())
 
 // Routers
 app.get('/', loginAuth, async (req, res) => {
-  var token = req.query["token"]
-
   try {
-    var tokenCreated = await Token.findOne({
+    var user = await User.findOne({
       where: {
-        token: token
-      },
-      include: {
-        model: User
+        id: req.user.id
       }})
 
-    if (tokenCreated) {
-      tokenCreated.user.password = '*******'
-
-      res.json(tokenCreated.user)
-    } else {
+    if (user) {
+      user.password = undefined
+      res.status(200)
       res.json({
-        error: 'Login inválido'
+        ...user.dataValues
+      })
+    } else {
+      res.status(404)
+      res.json({
+        error: 'Usuário não encontrado'
       })
     }
   } catch (error) {
+    res.status(500)
     res.json({
       error: 'Houve um erro inesperado'
     })
@@ -76,26 +75,22 @@ app.post('/', async (req, res) => {
 
       var hash = bcrypt.hashSync(password, bcrypt.genSaltSync(10))
 
-      var token_verify = await randomBytes(20)
+      var token_verify = await randomBytes(5)
       token_verify = token_verify.toString('hex')
 
-      var text1Email = `Muito obrigado por criar a sua conta na Mycroway! Por segurança precisamos que você clique no butão abaixo para fazer a verificação do seu email`
-      var hostEmail = req.protocol+'://'+req.headers.host+'/verify?token_verify='+token_verify
+      var text1Email = `Muito obrigado por criar a sua conta na Mycroway! Por segurança precisamos que você faça a verificação do seu email.`
 
-      var verifyEmail = new emailTemplate(name, text1Email, hostEmail, 'Verificar email', 'Caso você não tenha criado uma conta na Mycroway, não clique no butão de confirmação acima.')
+      var verifyEmail = new emailTemplate(name, text1Email, '/', '', 'Caso você não tenha criado uma conta na Mycroway, ignore este e-mail.', token_verify)
 
       var user = await User.create({
         name,
         email,
         gender,
         password: hash,
+        token: token_verify
       })
-      user.password = '*********'
-
-      var token = await TokenVerify.create({
-        token: token_verify,
-        userId: user.id
-      })
+      user.password = undefined
+      user.token = undefined
 
       emailConfig.transporter.sendMail({
         from: `Mycroway <${process.env.EMAIL}>`,
@@ -104,42 +99,39 @@ app.post('/', async (req, res) => {
         html: verifyEmail.Render()
 
       })
-
-      res.json(user)
+      res.status(200)
+      res.json({
+        user, msg: 'Enviamos um e-mail com um token de verificação.'
+      })
     }
   }
 })
 
 app.patch('/verify', loginAuth, async (req, res) => {
-  var token = req.query["token_verify"]
+  var token = req.query["token"]
 
-  var tokenCreated = await TokenVerify.findOne({
+  if (!token) {
+    res.status(401)
+    res.json({
+      error: 'Token inválido!'
+    })
+  }
+
+  var user = await User.findOne({
     where: {
       token: token
-    },
-    include: [{
-      model: User
-    }]
-  })
+    }})
 
-  if (tokenCreated && !tokenCreated.user.emailChecked) {
-    tokenCreated.user.password = '********'
+  if (user) {
+    user.password = undefined
 
     try {
-      User.update({
-        emailChecked: true
-      }, {
-        where: {
-          id: tokenCreated.user.id
-        }
+      user.update({
+        emailChecked: true,
+        token: null
       })
 
-      TokenVerify.destroy({
-        where: {
-          id: tokenCreated.id
-        }
-      })
-
+      res.status(200)
       res.json({
         msg: 'Perfeito! o seu e-mail foi verificado com sucesso!'
       })
@@ -156,7 +148,7 @@ app.patch('/verify', loginAuth, async (req, res) => {
   }
 })
 
-app.post('/login', async (req, res) => {
+app.post('/auth', async (req, res) => {
   var {
     email,
     password
@@ -172,59 +164,38 @@ app.post('/login', async (req, res) => {
       var correct = bcrypt.compareSync(password, user.password)
 
       if (correct) {
-
-        var token = await randomBytes(20)
-        token = token.toString('hex')
-
-        Token.create({
-          token,
+        jwt.sign({
           userId: user.id
+        }, JWTSecret, {
+          expiresIn: '12h'
+        }, (error, token) => {
+          if (error) {
+            res.status(500)
+            res.json({
+              error
+            })
+          } else {
+            res.status(200)
+            res.json({
+              token
+            })
+          }
         })
-
-        user.password = '*******'
-
-        res.json({
-          msg: 'Usuário logado com sucesso!', token, user
-        })
-
       } else {
+        res.status(401)
         res.json({
           error: 'Senha incorreta!'
         })
       }
 
     } else {
+      res.status(400)
       res.json({
         error: 'Alguns dos dados estavam incorretos'
       })
     }
   } catch (error) {
-    res.json({
-      error: 'Houve um erro inesperado!'
-    })
-  }
-})
-
-app.post('/logout', loginAuth, async (req, res) => {
-  var token = req.query["token"]
-
-  try {
-    Token.destroy({
-      where: {
-        token: token
-      }
-    }).then(() => {
-
-      res.json({
-        msg: 'O logout foi feito com sucesso!'
-      })
-
-    }).catch((error) => {
-      res.json({
-        error: 'Houve um erro inesperado'
-      })
-    })
-  } catch (error) {
+    res.status(500)
     res.json({
       error: 'Houve um erro inesperado!'
     })
@@ -232,62 +203,63 @@ app.post('/logout', loginAuth, async (req, res) => {
 })
 
 app.delete('/', loginAuth, async (req, res) => {
-  const token = req.query["token"]
-  const tokenDatas = await Token.findOne({
-    where: {
-      token: token
-    }})
-
   try {
-    User.destroy({
+    var user = await User.findOne({
       where: {
-        id: tokenDatas.userId
+        id: req.user.id
       }})
 
-    Token.destroy({
-      where: {
-        userId: tokenDatas.userId
-      }
-    })
-    res.json({
-      msg: 'Usuário excluído com sucesso!'
-    })
+    if (user) {
+      User.destroy({
+        where: {
+          id: req.user.id
+        }})
+
+      res.status(200)
+      res.json({
+        msg: 'Usuário excluído com sucesso!'
+      })
+    } else {
+      res.status(404)
+      res.json({
+        error: 'Usuário inválido!'
+      })
+    }
   } catch (error) {
+    res.status(500)
     res.json({
       error: 'Houve um erro ao excluir o usuário!'
     })
   }
 })
 
-app.patch('/update', loginAuth, async (req, res) => {
-  const token = req.query["token"]
-  const tokenDatas = await Token.findOne({
-    where: {
-      token: token
-    }
-  })
+app.patch('/', loginAuth, async (req, res) => {
   const datas = req.body
 
   if (datas.email) {
+    res.status(401)
     res.json({
       error: 'O email não pode ser atualizado!'
     })
   } else if (datas.password) {
+    res.status(401)
     res.json({
       error: 'Acesse a página de redefinição de senha para atualiza-la'
     })
   } else {
     try {
-      var newUser = await User.update({
+      User.update({
         ...datas
       }, {
         where: {
-          id: tokenDatas.userId
+          id: req.user.id
         }})
+      res.status(200)
       res.json({
         msg: 'Usuário atualizado com sucesso!'
       })
     } catch (error) {
+      res.status(500)
       res.json({
         error: 'Houve um erro inesperado'
       })
@@ -310,7 +282,7 @@ app.post('/forgot_password', async (req, res) => {
       token = token.toString('hex')
 
       User.update({
-        TokenReset: token
+        token: token
       }, {
         where: {
           id: user.id
@@ -345,32 +317,49 @@ app.post('/forgot_password', async (req, res) => {
 })
 
 app.patch('/reset', async (req, res) => {
-  var { token, password } = req.body
-  
+  var {
+    token,
+    password
+  } = req.body
+
   if (!token) {
     res.status(404)
-    res.json({error: 'Token inválido!'})
+    res.json({
+      error: 'Token inválido!'
+    })
   }
-  
+
   if (!password) {
     res.status(404)
-    res.json({error: 'Senha invalida!'})
+    res.json({
+      error: 'Senha invalida!'
+    })
   }
-  
-  var user = await User.findOne({where: {
-    TokenReset: token
-  }})
-  
+
+  var user = await User.findOne({
+    where: {
+      token: token
+    }})
+
   if (user) {
     var hash = bcrypt.hashSync(password, bcrypt.genSaltSync(10))
-    User.update({password: hash, TokenReset: null}, {where: {id: user.id}})
+    User.update({
+      password: hash, token: null
+    }, {
+      where: {
+        id: user.id
+      }})
     res.status(200)
-    res.json({msg: 'Senha redefinida com sucesso!'})
+    res.json({
+      msg: 'Senha redefinida com sucesso!'
+    })
   } else {
     res.status(404)
-    res.json({error: 'Token inválido!'})
+    res.json({
+      error: 'Token inválido!'
+    })
   }
-  
+
 })
